switched multimedia app to work with floats instead of doubles

+++ distributed/apps/multimedia/helperClasses.scala
@@ -1,5 +1,5 @@
 class FeatureVec extends Serializable{
 	var weight = 0
 	var tag = 0
-	var vecWeights = Array[Double]()
+	var vecWeights = Array[Float]()
 }
